Remove boilerplate HTML example from README
HTML is left out on GitHub other than in raw file mode. This caused the "e.g. example code here" to display "e.g. ," and continue with the list of README items which can be confusing if the comma is interpreted as an error rather than interpreting HTML as being edited out by GitHub.